DeutschJozsa

Description

The Deustch-Jozsa Oracle algorithm

Documentation

type Oracle = ([QBit], QBit) -> QM ([QBit], QBit)

balanced :: Oracle

An oracle with a balanced function

constant :: Oracle

An oracle with a constant function

deutschJozsa :: Int -> Oracle -> QM [Bit]

Will return a list of ones if balanced and list of zeros if constant. Size is the number of qubit inputs to the oracle.