Internal.Core

Description

Internal matrix and measurment operations

Documentation

appendState :: QState -> QState -> QState

Appends state with tensor

tensorVector :: Vector C -> Vector C -> Vector C

newVector :: Bit -> QState

Vector state representations of qubits with 100% probaility to collapse to their bit counterparts

type Prob = Rational

findQbitProb1 :: QBit -> QState -> Prob

Finds the probability of the qubit measing to a 1. Find all the amplitudes where that qubit is one and converts it to probabilities.

type Amplitude = Complex Double

findMarginAmps1 :: QBit -> QState -> [Amplitude]

Finds the amplitudes from all the positions where that qubit is one.

maskMatch :: Int -> Int -> Bool

If all the bits in b are in a it is a match.

qbitMask :: QBit -> QState -> Int

Given a qbit, finds its mask in the qstate. E.g. an amplitude in a 3 qbits state could be |100>. A mask of 100=8 is wanted if the zero'th qbit is in interest, same mask would work for |101>. If qbit is 1 we want it to be 010=4...

ampToProb :: Amplitude -> Prob

Given a complex amplitude, will return its probability.

remImpossibleStates :: QState -> QBit -> Bit -> QState

Removes the states that contradict the measurment from the qbit to bit, also normalizes the state to a length of one.

qstateAmps :: QState -> [(Ix, Amplitude)]

From a qstate, returns the amplitudes with its indexes.

rngQbit :: Prob -> IO Bit