Internal.Gates

Description

Internal matrix operations

Documentation

i :: Complex Double

The imaginary unit

applyParallel :: Matrix C -> Matrix C -> Matrix C

applyGate :: Matrix C -> QM ()

Apply gate to the current quantum state

insertAt :: a -> Int -> [a] -> [a]

Insert an element at a specific index in a list.

runGate :: Matrix C -> QBit -> QM QBit

Apply a 2x2 gate, to a specific qubit.

It will update the qstate.

parallel :: [Matrix C] -> QM ()

proj0 :: Matrix C

Projection of the zero basis vector

proj1 :: Matrix C

Projection of the one basis vector

controlMatrix :: Int -> QBit -> QBit -> Matrix C -> Matrix C

Produce matrix running a gate controlled by another bit

ccontrolMatrix :: Int -> QBit -> QBit -> QBit -> Matrix C -> Matrix C

Produce a matrix running a gate controlled by two other bits