QM

Description

Definition of the quantum monad (QM) and helper functions

The Quantum Monad

data QM a

The Quantum Monad Keeps a state of the complex vector representation while allowing pseudo-random number generation

run :: QM a -> IO a

Run quantum program, discarding it's final state

runDebug :: QM a -> IO a

Run quantum program, prints the final quantum state and returns the result

eval :: QM a -> IO (a, QState)

Run quantum program

put :: QState -> QM ()

Replace the quantum state

get :: QM QState

Fetch the quantum state

modify :: (QState -> QState) -> QM ()

Apply a function to the quantum state

io :: IO a -> QM a

Perform IO action inside the quantum monad

Quantum State representation

newtype QState

The program state, a complex vector representation of the qubits in the system

Constructors

QState

Fields

state :: Vector C

QBit

type Ix = Int

Internal index type, to indicate it is a qubit index.

newtype QBit

Pointer to a qubit in QState. Represents the linking function in QLambda

Constructors

Ptr

Fields

link :: Ix

Helpers

checkState :: QM ()

Print the quantum state during operation

getState :: QM (QState, Int)

Returns the quantum state together with it's size

stateSize :: QState -> Ix

Given a QState, returns how many qbits it consists of.