Internal.Gates

Description

Internal matrix operations

Documentation

i :: Complex Double

The imaginary unit

applyParallel :: Matrix C -> Matrix C -> Matrix C

applyGate :: Matrix C -> QM ()

Apply gate to the current quantum state

changeAt :: a -> Int -> [a] -> [a]

Changes an element at an index in a list. index 0 will change the first element.

runGate :: Matrix C -> QBit -> QM QBit

Apply a 2x2 gate, to a specific qubit.

It will update the qstate.

parallel :: Int -> [(Matrix C, QBit)] -> Matrix C

controlMatrix :: Int -> QBit -> QBit -> Matrix C -> Matrix C

Produce matrix running a gate controlled by another bit

ccontrolMatrix :: Int -> QBit -> QBit -> QBit -> Matrix C -> Matrix C

Produce a matrix running a gate controlled by two other bits

proj0 :: Matrix C

Projection of the zero basis vector

proj1 :: Matrix C

Projection of the one basis vector

hmat :: Matrix C

Hadamard matrix

cmat :: Matrix C

CNOT matrix

phasemat :: Double -> Matrix C

Generic phase matrix, takes in phase change as radians

pXmat :: Matrix C

PauliX matrix

pYmat :: Matrix C

PauliY matrix

pZmat :: Matrix C

PauliZ matrix

idmat :: Matrix C

Identity matrix

Orphan instances

Show (Matrix C)

Instance details

Methods

showsPrec :: Int -> Matrix C -> ShowS

show :: Matrix C -> String

showList :: [Matrix C] -> ShowS